+++ framework/src/Volo.Abp.Cli.Core/Volo/Abp/ProjectBuilding/Building/ProjectResult.cs
@@ -1,4 +1,4 @@
-﻿namespace Volo.Abp.SolutionTemplating.Building
+﻿namespace Volo.Abp.ProjectBuilding.Building
 {
     public class ProjectResult
     {
