+++ framework/src/Volo.Abp.Cli.Core/Volo/Abp/ProjectBuilding/Building/Steps/CreateProjectResultZipStep.cs
@@ -1,8 +1,8 @@
 ﻿using Ionic.Zip;
-using Volo.Abp.SolutionTemplating.Files;
-using Volo.Abp.SolutionTemplating.Zipping;
+using Volo.Abp.ProjectBuilding.Files;
+using Volo.Abp.ProjectBuilding.Zipping;
 
-namespace Volo.Abp.SolutionTemplating.Building.Steps
+namespace Volo.Abp.ProjectBuilding.Building.Steps
 {
     public class CreateProjectResultZipStep : ProjectBuildPipelineStep
     {
