+++ framework/src/Volo.Abp.Cli.Core/Volo/Abp/ProjectBuilding/Building/Steps/NugetReferenceReplaceStep.cs
@@ -2,9 +2,9 @@
 using System.IO;
 using System.Linq;
 using HtmlAgilityPack;
-using Volo.Abp.SolutionTemplating.Files;
+using Volo.Abp.ProjectBuilding.Files;
 
-namespace Volo.Abp.SolutionTemplating.Building.Steps
+namespace Volo.Abp.ProjectBuilding.Building.Steps
 {
     public class NugetReferenceReplaceStep : ProjectBuildPipelineStep
     {
