+++ framework/src/Volo.Abp.Cli.Core/Volo/Abp/ProjectBuilding/Building/Steps/TemplateCodeDeleteStep.cs
@@ -1,6 +1,6 @@
-﻿using Volo.Abp.SolutionTemplating.Files;
+﻿using Volo.Abp.ProjectBuilding.Files;
 
-namespace Volo.Abp.SolutionTemplating.Building.Steps
+namespace Volo.Abp.ProjectBuilding.Building.Steps
 {
     public class TemplateCodeDeleteStep : ProjectBuildPipelineStep
     {
