+++ framework/src/Volo.Abp.Cli.Core/Volo/Abp/ProjectBuilding/DownloadInfo.cs
@@ -1,9 +1,9 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using Volo.Abp.Domain.Entities.Auditing;
-using Volo.Abp.SolutionTemplating.Building;
+using Volo.Abp.ProjectBuilding.Building;
 
-namespace Volo.Abp.SolutionTemplating
+namespace Volo.Abp.ProjectBuilding
 {
     [Table("Downloads")]
     public class DownloadInfo : CreationAuditedAggregateRoot<int>
