+++ framework/src/Volo.Abp.Cli.Core/Volo/Abp/ProjectBuilding/Files/FileEntryExtensions.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Volo.Abp.SolutionTemplating.Files
+namespace Volo.Abp.ProjectBuilding.Files
 {
     public static class FileEntryExtensions
     {
