+++ framework/src/Volo.Abp.Cli.Core/Volo/Abp/ProjectBuilding/Files/FileEntryList.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Volo.Abp.SolutionTemplating.Files
+namespace Volo.Abp.ProjectBuilding.Files
 {
     public class FileEntryList : List<FileEntry>
     {
