+++ framework/src/Volo.Abp.Cli.Core/Volo/Abp/ProjectBuilding/Files/FileEntryListExtensions.cs
@@ -1,7 +1,7 @@
 using System;
 using Ionic.Zip;
 
-namespace Volo.Abp.SolutionTemplating.Files
+namespace Volo.Abp.ProjectBuilding.Files
 {
     public static class FileEntryListExtensions
     {
