+++ framework/src/Volo.Abp.Cli.Core/Volo/Abp/ProjectBuilding/ProjectBuildResult.cs
@@ -1,4 +1,4 @@
-﻿namespace Volo.Abp.SolutionTemplating
+﻿namespace Volo.Abp.ProjectBuilding
 {
     public class ProjectBuildResult
     {
