+++ framework/src/Volo.Abp.Cli.Core/Volo/Abp/ProjectBuilding/SolutionName.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Volo.Abp.SolutionTemplating
+namespace Volo.Abp.ProjectBuilding
 {
     public class SolutionName
     {
