+++ framework/src/Volo.Abp.Cli.Core/Volo/Abp/ProjectBuilding/Templates/ServiceTemplate.cs
@@ -1,4 +1,4 @@
-﻿using Volo.Abp.SolutionTemplating.Building;
+﻿using Volo.Abp.ProjectBuilding.Building;
 
 namespace Volo.Abp.ProjectBuilding.Templates
 {
