--- framework/src/Volo.Abp.Cli.Core/Volo/Abp/SolutionTemplating/StandardVersions.cs
@@ -1,9 +1,0 @@
-﻿namespace Volo.Abp.SolutionTemplating
-{
-    public static class StandardVersions
-    {
-        public const string LatestStable = "LatestStable";
-
-        public const string LatestUnstable = "LatestUnstable";
-    }
-}
